DDF-3497 Install Apache Ignite in DDF
This commit adds working installable Ignite features to DDF.

This is a PROTOTYPE. As of this commit, Ignite features are not ready to be added to a proper DDF release or master branch.

DDF should be largely unaffected by this initial change as no features currently utilize Ignite and Ignite features are set up for manual installation, not automatically installed on startup.